requirements.txt
fastapi
uvicorn==0.24.0
pydantic==2.7.2
motor==3.3.2
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
sentence-transformers>=2.2.2,<3
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
duckduckgo-search==6.2.13
huggingface_hub>=0.17.0,<1.0
transformers<5,>=4.41.2
tokenizers<0.20,>=0.19.1
groq==0.11.0  # Or latest: pip install groq --upgrade
python-multipart==0.0.9
faiss-cpu==1.7.4